Add types.ts for readability and refactor build
Some additional highlights:

- Add lightweight shim to encode string to uint8 array without window
- Switch build to commonjs
- Remove antipatterns from build
- Prefer yarn > npm
- Modify webpack package config
