+++ src/protocol.ts
@@ -13,10 +13,11 @@
  * limitations under the License.
  */
 
-import {Msg, NatsConnectionOptions, Payload, VERSION} from "./nats";
+import * as pkg from "../package.json";
+import {ClientHandlers, ConnectionOptions, Callback, ErrorCallback, Msg, Payload, Req, Sub} from "./types";
 import {Transport, TransportHandlers, WSTransport} from "./transport";
 import {ErrorCode, NatsError} from "./error";
-import {buildWSMessage, extend, extractProtocolMessage, settle} from "./util";
+import {buildWSMessage, extend, extractProtocolMessage, settle, stringToUint8Array} from "./util";
 import {Nuid} from "js-nuid"
 import {DataBuffer} from "./databuffer";
 
@@ -56,10 +57,10 @@ export class Connect {
     protocol: number = 1;
     user?: string;
     verbose: boolean = false;
-    version: string = VERSION;
+    version: string = pkg.version;
 
-    constructor(opts?: NatsConnectionOptions) {
-        opts = opts || {} as NatsConnectionOptions;
+    constructor(opts?: ConnectionOptions) {
+        opts = opts || {} as ConnectionOptions;
         if (opts.token) {
             this.auth_token = opts.token;
         }
@@ -77,46 +78,6 @@ export class Connect {
     }
 }
 
-export interface Callback {
-    (): void;
-}
-
-export interface ErrorCallback {
-    (error: Error): void;
-}
-
-export interface ClientHandlers {
-    closeHandler: Callback;
-    errorHandler: ErrorCallback;
-}
-
-
-export interface MsgCallback {
-    (msg: Msg): void;
-}
-
-export interface RequestOptions {
-    timeout?: number;
-}
-
-export interface Base {
-    subject: string;
-    callback: MsgCallback;
-    received: number;
-    timeout?: number | null;
-    max?: number | undefined;
-    draining: boolean;
-}
-
-export interface Req extends Base {
-    token: string;
-}
-
-export interface Sub extends Base {
-    sid: number;
-    queue?: string | null;
-}
-
 export function defaultSub(): Sub {
     return {sid: 0, subject: "", received: 0} as Sub;
 }
@@ -125,7 +86,6 @@ export function defaultReq(): Req {
     return {token: "", subject: "", received: 0, max: 1} as Req;
 
 }
-
 
 export class Request {
     token: string;
@@ -268,7 +228,6 @@ export class MuxSubscription {
     };
 }
 
-
 export class Subscriptions {
     mux!: Sub;
     subs: { [key: number]: Sub } = {};
@@ -320,7 +279,6 @@ export class Subscriptions {
     }
 }
 
-
 export class MsgBuffer {
     msg: Msg;
     length: number;
@@ -369,7 +327,7 @@ export class ProtocolHandler implements TransportHandlers {
     inbound: DataBuffer;
     infoReceived: boolean = false;
     muxSubscriptions: MuxSubscription;
-    options: NatsConnectionOptions;
+    options: ConnectionOptions;
     outbound: DataBuffer;
     payload: MsgBuffer | null = null;
     pongs: Array<Function | undefined> = [];
@@ -379,7 +337,7 @@ export class ProtocolHandler implements TransportHandlers {
     transport!: Transport;
     noMorePublishing: boolean = false;
 
-    constructor(options: NatsConnectionOptions, handlers: ClientHandlers) {
+    constructor(options: ConnectionOptions, handlers: ClientHandlers) {
         this.options = options;
         this.clientHandlers = handlers;
         this.subscriptions = new Subscriptions();
@@ -388,7 +346,7 @@ export class ProtocolHandler implements TransportHandlers {
         this.outbound = new DataBuffer();
     }
 
-    public static connect(options: NatsConnectionOptions, handlers: ClientHandlers): Promise<ProtocolHandler> {
+    public static connect(options: ConnectionOptions, handlers: ClientHandlers): Promise<ProtocolHandler> {
         return new Promise<ProtocolHandler>((resolve, reject) => {
             let ph = new ProtocolHandler(options, handlers);
             ph.connectError = reject;
@@ -545,7 +503,7 @@ export class ProtocolHandler implements TransportHandlers {
     sendCommand(cmd: string | ArrayBuffer) {
         let buf: ArrayBuffer;
         if (typeof cmd === 'string') {
-            buf = new TextEncoder().encode(cmd).buffer;
+            buf = stringToUint8Array(cmd);
         } else {
             buf = cmd as ArrayBuffer;
         }
