+++ test/properties.ts
@@ -13,18 +13,20 @@
  * limitations under the License.
  */
 
+import * as pkg from '../package.json';
 import test from "ava";
 import {Connect} from "../src/protocol";
-import {connect, NatsConnectionOptions, Payload, VERSION} from "../src/nats";
+import {connect} from "../src/nats";
+import {ConnectionOptions, Payload} from "../src/types";
 
 test('VERSION is semver', (t) => {
-    t.regex(VERSION, /[0-9]+\.[0-9]+\.[0-9]+/);
+    t.regex(pkg.version, /[0-9]+\.[0-9]+\.[0-9]+/);
 });
 
 test('VERSION matches package.json', (t) => {
     // we are getting build in lib/test
-    let pkg = require('../../package.json');
-    t.is(pkg.version, VERSION);
+    let pkgcontrol = require('../../package.json');
+    t.is(pkgcontrol.version, pkg.version);
 });
 
 test('connect is a function', (t) => {
@@ -45,7 +47,7 @@ test('default connect properties', (t) => {
 });
 
 test('configured options', (t) => {
-    let nco = {} as NatsConnectionOptions;
+    let nco = {} as ConnectionOptions;
     nco.payload = Payload.BINARY;
     nco.name = "test";
     nco.pass = "secret";
