+++ webpack.config.js
@@ -1,21 +1,31 @@
 const path = require('path')
+const webpack = require('webpack');
 
 module.exports = {
     mode: 'production',
-    entry: path.resolve(__dirname, 'src/nats.ts'),
+    entry: path.resolve(__dirname, 'src/index.ts'),
     output: {
         path: path.resolve(__dirname),
-        filename: 'nats.js',
-        libraryTarget: 'umd',
-        library: 'nats'
+        filename: 'dist/nats.js',
+        libraryTarget: 'commonjs',
+        library: 'nats',
     },
     resolve: {
-        extensions: ['.ts']
+        extensions: ['.ts'],
     },
     module: {
         rules: [
-            { test: /\.ts$/, loader: "ts-loader", exclude:  /(node_modules|test)/}
-        ]
+            {
+                test: /\.ts$/,
+                loader: "ts-loader",
+                exclude:  /(node_modules|test)/,
+            },
+        ],
     },
+    plugins: [
+        new webpack.DefinePlugin({
+            pkg: require("./package.json"),
+        }),
+    ],
     devtool: "source-map"
 };
